# SteticOpenApi.yaml
openapi: 3.0.1
info:
  title: Stetic API
  description: API para la gestión de un centro de estética
  version: 1.0.0
servers:
  - url: 'https://tiendasphy.000webhostapp.com/index.php'
...
paths:
...
  /employees:
    get:
      tags:
        - employees
      summary: Lista todos los empleados o filtra
      parameters:
        - name: name
          in: query
          description: Nombre del empleado
          required: false
          schema:
            type: string

        - name: dni
          in: query
          description: dni del empleado
          required: false
          schema:
            type: string

        - name: academicTitle
          in: query
          description: Si posee titulación
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Devuelve la lista completa de empleados o el empleado filtrado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
    post:
      tags:
        - employees
      summary: Añadir un nuevo empleado
      requestBody:
        description: Empleado a añadir
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          description: Empleado añadido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /employee/{employeeId}:
    get:
      tags:
        - employees
      summary: Obtiene un empleado por ID
      parameters:
        - name: employeeId
          in: path
          description: ID del empleado
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Devuelve el empleado solicitado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Empleado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - employees
      summary: Elimina un empleado por ID
      parameters:
        - name: employeeId
          in: path
          description: ID del empleado
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: El empleado se ha eliminado correctamente
        '404':
          description: Empleado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    put:
      tags:
        - employees
      summary: Modifica un empleado por ID
      parameters:
        - name: employeeId
          in: path
          description: ID del empleado
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: El empleado se ha modificado correctamente
        '404':
          description: Empleado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /shops:
    get:
      tags:
        - shops
      summary: Lista todas las tiendas o las filtradas
      parameters:
        - name: name
          in: query
          description: Nombre de la tienda
          required: false
          schema:
            type: string

        - name: city
          in: query
          description: ciudad de la tienda
          required: false
          schema:
            type: string

        - name: solarium
          in: query
          description: Si posee solarium
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Devuelve la lista completa de tiendas o las filtradas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shop'
    post:
      tags:
        - shops
      summary: Añadir una nueva tienda
      requestBody:
        description: Tienda a añadir
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shop'
      responses:
        '201':
          description: Tienda añadida correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /shop/{shopId}:
    get:
      tags:
        - shops
      summary: Obtiene una tienda por ID
      parameters:
        - name: shopId
          in: path
          description: ID de la tienda
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Devuelve la tienda solicitada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '404':
          description: Tienda no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - shops
      summary: Modificar una tienda por ID
      parameters:
        - name: shopId
          in: path
          description: ID de la tienda
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shop'
      responses:
        '200':
          description: La tienda se ha modificado correctamente
        '404':
          description: Tienda no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - shops
      summary: Eliminar una tienda por ID
      parameters:
        - name: shopId
          in: path
          description: ID de la tienda
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: La tienda se ha eliminado correctamente
        '404':
          description: Tienda no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        DNI:
          type: string
        age:
          type: integer
        height:
          type: number
          format: float
        academicTitle:
          type: boolean
        registerDate:
          type: string
          format: date
        shop:
          $ref: '#/components/schemas/Shop'
    Shop:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        employeesNumber:
          type: integer
        city:
          type: string
        meters:
          type: number
          format: float
        solarium:
          type: boolean
        openDate:
          type: string
          format: date
        employees:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: array
          items:
            type: string
    InternalServerError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...